requirements.txt
discord.py==2.6.4
python-dotenv==1.0.1
psutil==5.9.8
PyNaCl==1.5.0
yt-dlp>=2024.8.6
wavelink>=3.4.1
jishaku==2.5.2
